[ci] Build docker on change
This patch adds a CI job that builds a docker image
on every change in the docker folder or docker build
ci job. The job is triggered by the push event on
the main/develop branches only, so not verified
changes will not trigger the job and will not
upload the docker image to the registry.

Signed-off-by: Mykhailo Lohvynenko <Mykhailo_Lohvynenko@epam.com>